+++ packages/cli/src/modules/index.ts
@@ -1,2 +1,0 @@
-export * from "./publish-angular.js"
-export * from "./publish.js"